requirements.txt
crewai==0.152.0
crewai-tools==0.59.0
matplotlib==3.10.5
matplotlib-inline==0.1.7
pandas==2.3.1
plotly==6.2.0
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
streamlit==1.48.0
types-requests==2.32.4.20250611
